+++ src/main/java/io/getstream/models/AddReactionRequest.java
@@ -38,6 +38,10 @@ public class AddReactionRequest {
   private Boolean createNotificationActivity;
 
   @Nullable
+  @JsonProperty("create_users")
+  private Boolean createUsers;
+
+  @Nullable
   @JsonProperty("enforce_unique")
   private Boolean enforceUnique;